+++ src/lib/layout/search/index.tsx
@@ -251,7 +251,7 @@ export const SearchComponent = () => {
                   size="lg"
                   placeholder="Enter your keyword..."
                   style={{ maxHeight: "54px", border: "none" }}
-                  pr={28}
+                  pr={40}
                   pl={10}
                   value={keyword}
                   onChange={handleSearchChange}