+++ src/lib/pages/nft-collection-details/components/tables/activities/index.ts
@@ -1,0 +1,2 @@
+export * from "./ActivitiesFull";
+export * from "./ActivitiesSequencer";