+++ src/lib/services/nft/index.ts
@@ -1,5 +1,6 @@
 import type { UseQueryOptions } from "@tanstack/react-query";
-import { useQuery } from "@tanstack/react-query";
+import { useInfiniteQuery, useQuery } from "@tanstack/react-query";
+import { useMemo } from "react";
 
 import type {
   Metadata,
@@ -35,6 +36,7 @@ import {
   getNftByNftAddressSequencer,
   getNftMintInfoSequencer,
   getNftsByAccountSequencer,
+  getNftsSequencer,
   getNftTransactionsSequencer,
 } from "./sequencer";
 
@@ -42,7 +44,8 @@ export const useNfts = (
   collectionAddress: HexAddr32,
   limit: number,
   offset: number,
-  search = ""
+  search = "",
+  enabled = true
 ) => {
   const { chainConfig } = useCelatoneApp();
 
@@ -60,8 +63,44 @@ export const useNfts = (
     {
       retry: 1,
       refetchOnWindowFocus: false,
+      enabled,
     }
   );
+};
+
+export const useNftsSequencer = (
+  collectionAddress: HexAddr32,
+  limit: number,
+  offset: number,
+  search = "",
+  enabled = true
+) => {
+  const lcdEndpoint = useLcdEndpoint();
+
+  const { data, ...rest } = useQuery<Nft[]>(
+    [CELATONE_QUERY_KEYS.NFTS_SEQUENCER, lcdEndpoint, collectionAddress],
+    async () => getNftsSequencer(lcdEndpoint, collectionAddress),
+    {
+      retry: 1,
+      refetchOnWindowFocus: false,
+      enabled,
+    }
+  );
+
+  const computedData = useMemo(() => {
+    if (!data) return data;
+    const filteredData = data.filter(
+      (val) =>
+        val.tokenId.toLowerCase().includes(search.toLowerCase()) ||
+        val.nftAddress.toLowerCase() === search.toLowerCase()
+    );
+    return limit ? filteredData?.slice(offset, limit + offset) : filteredData;
+  }, [data, limit, offset, search]);
+
+  return {
+    data: computedData,
+    ...rest,
+  };
 };
 
 export const useNftByNftAddress = (
@@ -175,20 +214,22 @@ export const useNftTransactionsSequencer = (
 ) => {
   const lcdEndpoint = useLcdEndpoint();
 
-  return useQuery<NftTransactions[]>(
-    [
-      CELATONE_QUERY_KEYS.NFT_TRANSACTIONS_SEQUENCER,
-      lcdEndpoint,
-      nftAddress,
-      enabled,
-    ],
-    async () => getNftTransactionsSequencer(lcdEndpoint, nftAddress),
+  const { data, ...rest } = useInfiniteQuery(
+    [CELATONE_QUERY_KEYS.NFT_TRANSACTIONS_SEQUENCER, lcdEndpoint],
+    async ({ pageParam }) =>
+      getNftTransactionsSequencer(lcdEndpoint, pageParam, nftAddress),
     {
-      retry: 1,
+      getNextPageParam: (lastPage) => lastPage.pagination.nextKey ?? undefined,
       refetchOnWindowFocus: false,
       enabled,
+      retry: 1,
     }
   );
+
+  return {
+    data: data?.pages.flatMap((page) => page.items),
+    ...rest,
+  };
 };
 
 export const useNftTransactionsCount = (